NODE-180: start work to consistently treat Modbus register data as short values to match 16-bit size of Modbus words.

+++ net.solarnetwork.node.io.modbus.test/src/net/solarnetwork/node/io/modbus/test/ModbusDataUtilsTests.java
@@ -44,14 +44,14 @@ public class ModbusDataUtilsTests {
 
 	@Test
 	public void testParseFloat32() {
-		Float result = ModbusDataUtils.parseFloat32(0x403F, 0xA7F6);
+		Float result = ModbusDataUtils.parseFloat32((short) 0x403F, (short) 0xA7F6);
 		assertThat(result, notNullValue());
 		assertThat("Float value", result.doubleValue(), closeTo(2.994626, 0.000001));
 	}
 
 	@Test
 	public void testParseFloat32NaN() {
-		Float result = ModbusDataUtils.parseFloat32(0xFFC0, 0x0000);
+		Float result = ModbusDataUtils.parseFloat32((short) 0xFFC0, (short) 0x0000);
 		assertThat("Float value is NaN", result, nullValue());
 	}
 
@@ -70,7 +70,8 @@ public class ModbusDataUtilsTests {
 
 	@Test
 	public void testParseInt32() {
-		assertThat(ModbusDataUtils.parseInt32(0xFF44, 0x1C6F), Matchers.equalTo(-12313489));
+		assertThat(ModbusDataUtils.parseInt32((short) 0xFF44, (short) 0x1C6F),
+				Matchers.equalTo(-12313489));
 	}
 
 	@Test
@@ -173,8 +174,8 @@ public class ModbusDataUtilsTests {
 
 	@Test
 	public void testParseUnsignedIntegerWayLargerThan64() {
-		int[] words = new int[] { 0x0032, 0x8586, 0x616F, 0x5866, 0xFF00, 0x1755, 0x8668, 0x16FE, 0x2F58,
-				0x6685, 0x866B, 0x5866 };
+		short[] words = new short[] { (short) 0x0032, (short) 0x8586, 0x616F, 0x5866, (short) 0xFF00,
+				0x1755, (short) 0x8668, 0x16FE, 0x2F58, 0x6685, (short) 0x866B, 0x5866 };
 		BigInteger bint = ModbusDataUtils.parseUnsignedInteger(words, 0);
 		assertThat(bint,
 				equalTo(new BigInteger("00328586616F5866FF001755866816FE2F586685866B5866", 16)));
@@ -182,8 +183,8 @@ public class ModbusDataUtilsTests {
 
 	@Test
 	public void testParseUnsignedIntegerWayLargerThan64LeastSignificantWordOrder() {
-		int[] words = new int[] { 0x5866, 0x866B, 0x6685, 0x2F58, 0x16FE, 0x8668, 0x1755, 0xFF00, 0x5866,
-				0x616F, 0x8586, 0x0032 };
+		short[] words = new short[] { 0x5866, (short) 0x866B, 0x6685, 0x2F58, 0x16FE, (short) 0x8668,
+				0x1755, (short) 0xFF00, 0x5866, 0x616F, (short) 0x8586, 0x0032 };
 		BigInteger bint = ModbusDataUtils.parseUnsignedInteger(words, 0, LeastToMostSignificant);
 		assertThat(bint,
 				equalTo(new BigInteger("00328586616F5866FF001755866816FE2F586685866B5866", 16)));
@@ -191,8 +192,8 @@ public class ModbusDataUtilsTests {
 
 	@Test
 	public void testParseUnsignedIntegerWayLargerThan64NoLeadingZeros() {
-		int[] words = new int[] { 0xDC32, 0x8586, 0x616F, 0x5866, 0xFF00, 0x1755, 0x8668, 0x16FE, 0x2F58,
-				0x6685, 0x866B, 0x5866 };
+		short[] words = new short[] { (short) 0xDC32, (short) 0x8586, 0x616F, 0x5866, (short) 0xFF00,
+				0x1755, (short) 0x8668, 0x16FE, 0x2F58, 0x6685, (short) 0x866B, 0x5866 };
 		BigInteger bint = ModbusDataUtils.parseUnsignedInteger(words, 0);
 		assertThat(bint,
 				equalTo(new BigInteger("DC328586616F5866FF001755866816FE2F586685866B5866", 16)));
@@ -200,8 +201,8 @@ public class ModbusDataUtilsTests {
 
 	@Test
 	public void testParseUnsignedIntegerWayLargerThan64NoLeadingZerosLeastSignificantWordOrder() {
-		int[] words = new int[] { 0x5866, 0x866B, 0x6685, 0x2F58, 0x16FE, 0x8668, 0x1755, 0xFF00, 0x5866,
-				0x616F, 0x8586, 0xDC32 };
+		short[] words = new short[] { 0x5866, (short) 0x866B, 0x6685, 0x2F58, 0x16FE, (short) 0x8668,
+				0x1755, (short) 0xFF00, 0x5866, 0x616F, (short) 0x8586, (short) 0xDC32 };
 		BigInteger bint = ModbusDataUtils.parseUnsignedInteger(words, 0, LeastToMostSignificant);
 		assertThat(bint,
 				equalTo(new BigInteger("DC328586616F5866FF001755866816FE2F586685866B5866", 16)));
@@ -254,7 +255,7 @@ public class ModbusDataUtilsTests {
 
 	@Test
 	public void testParseBytes() {
-		int[] words = new int[] { 0x0103, 0x0507, 0x090b, 0x0d00 };
+		short[] words = new short[] { 0x0103, 0x0507, 0x090b, 0x0d00 };
 		byte[] data = ModbusDataUtils.parseBytes(words, 0);
 		Byte[] bytes = new Byte[data.length];
 		for ( int i = 0; i < bytes.length; i++ ) {
@@ -266,7 +267,7 @@ public class ModbusDataUtilsTests {
 
 	@Test
 	public void testParseBytesLeastToMostSignificant() {
-		int[] words = new int[] { 0x0d00, 0x090b, 0x0507, 0x0103 };
+		short[] words = new short[] { 0x0d00, 0x090b, 0x0507, 0x0103 };
 		byte[] data = ModbusDataUtils.parseBytes(words, 0, LeastToMostSignificant);
 		Byte[] bytes = new Byte[data.length];
 		for ( int i = 0; i < bytes.length; i++ ) {
